requirements.txt
django-crispy-forms~=1.6.0
django-environ~=0.4.0
django-filter~=0.13.0
django-guardian~=1.4.4
djangorestframework~=3.3.3
Django~=1.9.5
flake8-commas~=0.1.6
flake8~=2.5.4
isort~=4.2.5
Markdown~=2.6.6
psycopg2~=2.6.1
